Move ease function stuffs into ease_function module
