+++ src/curve/ease_function.rs
@@ -1,4 +1,828 @@
-use std::f32::consts::PI;
+use super::{AToB, CurveValue};
+use crate::TweenSystemSet;
+use bevy::prelude::*;
+use bevy_time_runner::TimeSpanProgress;
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+use std::f32::consts::{PI, TAU};
+
+/// Plugin for [`EaseFunction`]
+pub struct EaseFunctionPlugin;
+
+impl Plugin for EaseFunctionPlugin {
+    /// # Panics
+    ///
+    /// Panics if [`TweenAppResource`] does not exist in world.
+    ///
+    /// [`TweenAppResource`]: crate::TweenAppResource
+    fn build(&self, app: &mut App) {
+        let app_resource = app
+            .world()
+            .get_resource::<crate::TweenAppResource>()
+            .expect("`TweenAppResource` resource doesn't exist");
+        app.add_systems(
+            app_resource.schedule,
+            (
+                ease_function_a_to_b_system::<f32>(|&a, &b, &v| a.lerp(b, v)),
+                ease_function_a_to_b_system::<Vec2>(|&a, &b, &v| a.lerp(b, v)),
+                ease_function_a_to_b_system::<Vec3>(|&a, &b, &v| a.lerp(b, v)),
+                ease_function_a_to_b_system::<Quat>(|&a, &b, &v| a.slerp(b, v)),
+                ease_function_a_to_b_system::<Color>(|&a, b, &v| a.mix(b, v)),
+            )
+                .in_set(TweenSystemSet::UpdateCurveValue),
+        )
+        .register_type::<EaseFunction>();
+    }
+}
+
+// i'm very proud of this text art i made.
+// it's some what better than images because we can see them in terminal editors too
+// helix, vim, etc.
+
+/// Robert Penner's easing functions in an enum
+#[allow(missing_docs)]
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Component, Reflect)]
+#[reflect(Component)]
+#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
+pub enum EaseFunction {
+    #[default]
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗  
+    ///  │                                    ▄▀▘  
+    ///  │                                 ▄▞▀  ┆  
+    ///  │                              ▗▄▀     ┆  
+    ///  │                            ▄▀▘       ┆  
+    ///  │                         ▄▞▀          ┆  
+    ///  │                      ▗▄▀             ┆  
+    ///  │                    ▄▀▘               ┆  
+    ///  │                 ▄▞▀                  ┆  
+    ///  │              ▗▄▀                     ┆  
+    ///  │            ▄▀▘                       ┆  
+    ///  │         ▄▞▀                          ┆  
+    ///  │      ▗▄▀                             ┆  
+    ///  │    ▄▀▘                               ┆  
+    ///  │ ▄▞▀                                  ┆  
+    ///  ▄▀─────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    Linear,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗▀  
+    ///  │                                    ▄▘┆  
+    ///  │                                  ▗▞  ┆  
+    ///  │                                 ▄▘   ┆  
+    ///  │                               ▗▞     ┆  
+    ///  │                              ▞▘      ┆  
+    ///  │                            ▄▀        ┆  
+    ///  │                          ▄▀          ┆  
+    ///  │                        ▄▀            ┆  
+    ///  │                      ▄▀              ┆  
+    ///  │                   ▄▞▀                ┆  
+    ///  │                ▗▄▀                   ┆  
+    ///  │             ▄▞▀▘                     ┆  
+    ///  │        ▄▄▄▀▀                         ┆  
+    ///  ▄▄▄▄▄▞▀▀▀──────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuadraticIn,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄  
+    ///  │                           ▗▄▄▞▀▀▀    ┆  
+    ///  │                        ▄▞▀▘          ┆  
+    ///  │                    ▗▄▀▀              ┆  
+    ///  │                  ▄▀▘                 ┆  
+    ///  │               ▗▞▀                    ┆  
+    ///  │             ▗▞▘                      ┆  
+    ///  │           ▗▞▘                        ┆  
+    ///  │         ▗▞▘                          ┆  
+    ///  │        ▞▘                            ┆  
+    ///  │      ▄▀                              ┆  
+    ///  │    ▗▞                                ┆  
+    ///  │   ▄▘                                 ┆  
+    ///  │ ▗▞                                   ┆  
+    ///  │▗▘                                    ┆  
+    ///  ▞▘─────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuadraticOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗▄▄▄  
+    ///  │                               ▄▄▀▀▘  ┆  
+    ///  │                            ▗▞▀       ┆  
+    ///  │                          ▄▀▘         ┆  
+    ///  │                        ▄▀            ┆  
+    ///  │                      ▗▞              ┆  
+    ///  │                     ▞▘               ┆  
+    ///  │                   ▗▞                 ┆  
+    ///  │                  ▄▘                  ┆  
+    ///  │                 ▞                    ┆  
+    ///  │               ▄▀                     ┆  
+    ///  │             ▗▞                       ┆  
+    ///  │           ▗▞▘                        ┆  
+    ///  │         ▄▀▘                          ┆  
+    ///  │     ▗▄▞▀                             ┆  
+    ///  ▄▄▄▄▀▀▘────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuadraticInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▞  
+    ///  │                                     ▞┆  
+    ///  │                                    ▞ ┆  
+    ///  │                                   ▞  ┆  
+    ///  │                                  ▞   ┆  
+    ///  │                                ▗▞    ┆  
+    ///  │                               ▗▘     ┆  
+    ///  │                              ▞▘      ┆  
+    ///  │                            ▗▀        ┆  
+    ///  │                          ▗▞▘         ┆  
+    ///  │                        ▗▞▘           ┆  
+    ///  │                      ▄▞▘             ┆  
+    ///  │                   ▄▞▀                ┆  
+    ///  │              ▄▄▞▀▀                   ┆  
+    ///  ▄▄▄▄▄▄▄▄▄▄▞▀▀▀▀────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    CubicIn,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄▄▄▄▄▄  
+    ///  │                      ▄▄▞▀▀▀▀         ┆  
+    ///  │                  ▄▞▀▀                ┆  
+    ///  │               ▄▞▀                    ┆  
+    ///  │             ▄▀                       ┆  
+    ///  │           ▄▀                         ┆  
+    ///  │         ▗▀                           ┆  
+    ///  │        ▞▘                            ┆  
+    ///  │      ▗▀                              ┆  
+    ///  │     ▄▘                               ┆  
+    ///  │    ▞                                 ┆  
+    ///  │   ▞                                  ┆  
+    ///  │  ▞                                   ┆  
+    ///  │ ▞                                    ┆  
+    ///  │▞                                     ┆  
+    ///  ▞──────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    CubicOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄▄  
+    ///  │                            ▄▄▀▀▀     ┆  
+    ///  │                          ▄▀          ┆  
+    ///  │                        ▄▀            ┆  
+    ///  │                      ▗▞              ┆  
+    ///  │                     ▗▘               ┆  
+    ///  │                    ▗▘                ┆  
+    ///  │                   ▗▘                 ┆  
+    ///  │                  ▗▘                  ┆  
+    ///  │                 ▗▘                   ┆  
+    ///  │                ▗▘                    ┆  
+    ///  │               ▄▘                     ┆  
+    ///  │             ▗▞                       ┆  
+    ///  │           ▗▞▘                        ┆  
+    ///  │        ▗▄▞▘                          ┆  
+    ///  ▄▄▄▄▄▄▞▀▀▘─────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    CubicInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▐  
+    ///  │                                     ▗▘  
+    ///  │                                     ▞┆  
+    ///  │                                    ▞ ┆  
+    ///  │                                   ▞  ┆  
+    ///  │                                  ▞   ┆  
+    ///  │                                 ▞    ┆  
+    ///  │                                ▞     ┆  
+    ///  │                              ▗▞      ┆  
+    ///  │                             ▗▘       ┆  
+    ///  │                           ▗▞▘        ┆  
+    ///  │                         ▗▞▘          ┆  
+    ///  │                      ▗▄▀▘            ┆  
+    ///  │                  ▄▄▞▀▘               ┆  
+    ///  ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▞▀▀▀▀────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuarticIn,
+    /// ```txt
+    ///  ▲                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄▄▄▄▄▄▄▄▄▄  
+    ///  │                  ▄▄▞▀▀▀▀             ┆  
+    ///  │              ▗▄▀▀                    ┆  
+    ///  │            ▄▀▘                       ┆  
+    ///  │          ▄▀                          ┆  
+    ///  │        ▗▀                            ┆  
+    ///  │       ▄▘                             ┆  
+    ///  │      ▞                               ┆  
+    ///  │     ▞                                ┆  
+    ///  │    ▞                                 ┆  
+    ///  │   ▞                                  ┆  
+    ///  │  ▞                                   ┆  
+    ///  │ ▞                                    ┆  
+    ///  │▗▘                                    ┆  
+    ///  │▌                                     ┆  
+    ///  ▞──────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuarticOut,
+    /// ```txt
+    ///  ▲                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗▄▄▄▄▄▄▄▄  
+    ///  │                          ▗▄▀▀▘       ┆  
+    ///  │                        ▗▞▘           ┆  
+    ///  │                       ▞▘             ┆  
+    ///  │                      ▞               ┆  
+    ///  │                     ▞                ┆  
+    ///  │                    ▞                 ┆  
+    ///  │                   ▗▘                 ┆  
+    ///  │                  ▗▘                  ┆  
+    ///  │                  ▞                   ┆  
+    ///  │                 ▞                    ┆  
+    ///  │                ▞                     ┆  
+    ///  │               ▞                      ┆  
+    ///  │             ▄▀                       ┆  
+    ///  │          ▗▄▀                         ┆  
+    ///  ▄▄▄▄▄▄▄▄▄▀▀▘───────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuarticInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗  
+    ///  │                                      ▞  
+    ///  │                                     ▐┆  
+    ///  │                                    ▗▘┆  
+    ///  │                                    ▞ ┆  
+    ///  │                                   ▞  ┆  
+    ///  │                                  ▐   ┆  
+    ///  │                                 ▗▘   ┆  
+    ///  │                                ▗▘    ┆  
+    ///  │                               ▞▘     ┆  
+    ///  │                             ▗▞       ┆  
+    ///  │                           ▗▞▘        ┆  
+    ///  │                         ▗▞▘          ┆  
+    ///  │                     ▄▄▞▀▘            ┆  
+    ///  ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▀▀▀▀─────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuinticIn,
+    /// ```txt
+    ///  ▲                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄  
+    ///  │               ▄▄▞▀▀▀▘                ┆  
+    ///  │            ▄▀▀                       ┆  
+    ///  │          ▄▀                          ┆  
+    ///  │        ▄▀                            ┆  
+    ///  │       ▞                              ┆  
+    ///  │     ▗▀                               ┆  
+    ///  │    ▗▘                                ┆  
+    ///  │    ▌                                 ┆  
+    ///  │   ▞                                  ┆  
+    ///  │  ▞                                   ┆  
+    ///  │ ▗▘                                   ┆  
+    ///  │ ▌                                    ┆  
+    ///  │▞                                     ┆  
+    ///  ▗▘                                     ┆  
+    ///  ▞──────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuinticOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄▄▄▄▄▄  
+    ///  │                         ▗▞▀▀┄┄┄┄┄┄┄┄┄┄  
+    ///  │                       ▗▞▘            ┆  
+    ///  │                      ▗▘              ┆  
+    ///  │                     ▗▘               ┆  
+    ///  │                    ▗▘                ┆  
+    ///  │                    ▞                 ┆  
+    ///  │                   ▐                  ┆  
+    ///  │                   ▌                  ┆  
+    ///  │                  ▞                   ┆  
+    ///  │                 ▗▘                   ┆  
+    ///  │                ▗▘                    ┆  
+    ///  │               ▗▘                     ┆  
+    ///  │              ▄▘                      ┆  
+    ///  │            ▄▀                        ┆  
+    ///  ▄▄▄▄▄▄▄▄▄▄▞▀▀──────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    QuinticInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▀  
+    ///  │                                   ▗▞ ┆  
+    ///  │                                 ▗▞▘  ┆  
+    ///  │                                ▄▘    ┆  
+    ///  │                              ▄▀      ┆  
+    ///  │                            ▗▀        ┆  
+    ///  │                          ▗▞▘         ┆  
+    ///  │                        ▗▞▘           ┆  
+    ///  │                      ▗▞▘             ┆  
+    ///  │                    ▄▀▘               ┆  
+    ///  │                 ▗▄▀                  ┆  
+    ///  │               ▄▞▘                    ┆  
+    ///  │           ▗▄▞▀                       ┆  
+    ///  │       ▗▄▞▀▘                          ┆  
+    ///  ▄▄▄▄▄▀▀▀▘──────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    SineIn,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗▄▄▄▄  
+    ///  │                             ▄▄▀▀▀▘   ┆  
+    ///  │                         ▄▄▀▀         ┆  
+    ///  │                      ▄▞▀             ┆  
+    ///  │                   ▗▄▀                ┆  
+    ///  │                 ▗▞▘                  ┆  
+    ///  │               ▄▀▘                    ┆  
+    ///  │             ▄▀                       ┆  
+    ///  │           ▄▀                         ┆  
+    ///  │         ▗▀                           ┆  
+    ///  │       ▗▞▘                            ┆  
+    ///  │     ▗▞▘                              ┆  
+    ///  │    ▄▘                                ┆  
+    ///  │  ▄▀                                  ┆  
+    ///  │▗▞                                    ┆  
+    ///  ▞▘─────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    SineOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄  
+    ///  │                                ▄▞▀▀▘ ┆  
+    ///  │                             ▄▞▀      ┆  
+    ///  │                           ▄▀         ┆  
+    ///  │                         ▄▀           ┆  
+    ///  │                       ▄▀             ┆  
+    ///  │                     ▗▀               ┆  
+    ///  │                   ▗▞▘                ┆  
+    ///  │                  ▄▘                  ┆  
+    ///  │                ▗▀                    ┆  
+    ///  │              ▗▞▘                     ┆  
+    ///  │            ▗▞▘                       ┆  
+    ///  │          ▗▞▘                         ┆  
+    ///  │        ▄▞▘                           ┆  
+    ///  │     ▄▞▀                              ┆  
+    ///  ▄▄▄▞▀▀─────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    SineInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄  
+    ///  │                                      ▐  
+    ///  │                                     ▗▘  
+    ///  │                                    ▗▘┆  
+    ///  │                                   ▗▘ ┆  
+    ///  │                                  ▗▘  ┆  
+    ///  │                                ▗▞▘   ┆  
+    ///  │                              ▗▞▘     ┆  
+    ///  │                            ▗▞▘       ┆  
+    ///  │                         ▗▄▀▘         ┆  
+    ///  │                      ▗▄▀▘            ┆  
+    ///  │                  ▄▄▞▀▘               ┆  
+    ///  │           ▗▄▄▄▀▀▀                    ┆  
+    ///  ▄▄▄▄▄▄▄▞▀▀▀▀▘──────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    CircularIn,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄▄▄  
+    ///  │                       ▗▄▄▄▀▀▀▀▀▘     ┆  
+    ///  │                  ▄▄▞▀▀▘              ┆  
+    ///  │              ▗▄▀▀                    ┆  
+    ///  │           ▗▄▀▘                       ┆  
+    ///  │         ▄▀▘                          ┆  
+    ///  │       ▄▀                             ┆  
+    ///  │     ▄▀                               ┆  
+    ///  │   ▗▀                                 ┆  
+    ///  │  ▗▘                                  ┆  
+    ///  │ ▗▘                                   ┆  
+    ///  │▗▘                                    ┆  
+    ///  │▌                                     ┆  
+    ///  ▐                                      ┆  
+    ///  ▐                                      ┆  
+    ///  ▞──────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    CircularOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄  
+    ///  │                            ▄▄▞▀▀▀▘   ┆  
+    ///  │                         ▄▞▀          ┆  
+    ///  │                       ▄▀             ┆  
+    ///  │                     ▄▀               ┆  
+    ///  │                    ▞                 ┆  
+    ///  │                   ▗▘                 ┆  
+    ///  │                   ▐                  ┆  
+    ///  │                   ▌                  ┆  
+    ///  │                  ▗▘                  ┆  
+    ///  │                  ▞                   ┆  
+    ///  │                ▗▞                    ┆  
+    ///  │              ▗▞▘                     ┆  
+    ///  │            ▄▞▘                       ┆  
+    ///  │        ▄▄▞▀                          ┆  
+    ///  ▄▄▄▄▄▞▀▀▀──────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    CircularInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗  
+    ///  │                                      ▐  
+    ///  │                                      ▌  
+    ///  │                                     ▐┆  
+    ///  │                                     ▌┆  
+    ///  │                                    ▞ ┆  
+    ///  │                                   ▗▘ ┆  
+    ///  │                                  ▗▘  ┆  
+    ///  │                                 ▗▘   ┆  
+    ///  │                                ▗▘    ┆  
+    ///  │                               ▄▘     ┆  
+    ///  │                             ▗▞       ┆  
+    ///  │                           ▄▞▘        ┆  
+    ///  │                      ▄▄▄▀▀           ┆  
+    ///  ▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▞▀▀▀▀▀▀────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    ExponentialIn,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄  
+    ///  │             ▗▄▄▞▀▀▀▀▀▀▘              ┆  
+    ///  │          ▄▞▀▘                        ┆  
+    ///  │        ▄▀                            ┆  
+    ///  │      ▗▀                              ┆  
+    ///  │     ▗▘                               ┆  
+    ///  │    ▗▘                                ┆  
+    ///  │   ▗▘                                 ┆  
+    ///  │  ▗▘                                  ┆  
+    ///  │  ▞                                   ┆  
+    ///  │ ▐                                    ┆  
+    ///  │ ▌                                    ┆  
+    ///  │▐                                     ┆  
+    ///  │▌                                     ┆  
+    ///  ▗▘                                     ┆  
+    ///  ▞──────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    ExponentialOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▄▄▄▄▄▄▄▄▄▄  
+    ///  │                        ▗▄▀▀▀▘        ┆  
+    ///  │                      ▗▞▘             ┆  
+    ///  │                     ▗▘               ┆  
+    ///  │                     ▌                ┆  
+    ///  │                    ▐                 ┆  
+    ///  │                    ▌                 ┆  
+    ///  │                   ▐                  ┆  
+    ///  │                   ▌                  ┆  
+    ///  │                  ▐                   ┆  
+    ///  │                  ▌                   ┆  
+    ///  │                 ▞                    ┆  
+    ///  │                ▗▘                    ┆  
+    ///  │               ▄▘                     ┆  
+    ///  │            ▗▄▀                       ┆  
+    ///  ▄▄▄▄▄▄▄▄▄▄▞▀▀▘─────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    ExponentialInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄  
+    ///  │                                      ┆  
+    ///  │                                      ┆  
+    ///  │                                      ┆  
+    ///  │                                      ┆  
+    ///  │                                    ▐▖┆  
+    ///  │                                    ▐▐┆  
+    ///  │                                    ▐▐┆  
+    ///  │                                 ▗  ▐▐┆  
+    ///  │                                 ▐▌ ▐▐┆  
+    ///  │                              ▗  ▐▐ ▐▐┆  
+    ///  │                              ▐▌ ▐▐ ▐▐┆  
+    ///  │                           ▐▖ ▞▚ ▞▐ ▞▐┆  
+    ///  │                  ▄  ▄  ▛▖ ▌▚ ▌▐ ▌▐ ▌▐┆  
+    ///  ▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▀▚▞─▚▞─▚▐─▐▗▘▐─▌▐─▌▐─▌▐┬─▶
+    /// 0                        ▘ ▝▞ ▐▗▘▐ ▌▐ ▌▝▖  
+    ///                                ▛  ▌▌ ▌▌ ▌  
+    ///                                   █  ▌▌ ▌  
+    ///                                   ▛  ▌▌ ▌  
+    ///                                      ▌▌ ▌  
+    ///                                      █  ▌  
+    ///                                      ▘  ▌  
+    ///                                         ▌  
+    ///                                         ▌  
+    ///                                         ▌  
+    ///                                         ▚  
+    /// ```
+    ElasticIn,
+    /// ```txt
+    ///    ▗                                       
+    ///    ▌▌                                      
+    ///    ▌▌                                      
+    ///    ▌▌                                      
+    ///    ▌▌ ▐▖                                   
+    ///    ▌▌ ▌▌                                   
+    ///    ▌▌ ▌▌ ▐▖                                
+    ///  ▲ ▌▌ ▌▌ ▞▌ ▗                              
+    ///  │ ▌▚ ▌▚ ▌▚ ▞▌ ▗▖                          
+    /// 1┤┄▌▐┄▌▐┄▌▐┄▌▐┄▌▐┄▞▚┄▞▚┄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄  
+    ///  ▐▗▘▐ ▌▐ ▌▐ ▌▝▄▘ ▛  ▀  ▀                ┆  
+    ///  ▐▐ ▐▐ ▐▐ ▝▟  ▘                         ┆  
+    ///  ▐▐ ▐▐ ▐▐  ▘                            ┆  
+    ///  ▐▐ ▐▐  █                               ┆  
+    ///  ▐▐ ▐▐  ▘                               ┆  
+    ///  ▐▐ ▐▐                                  ┆  
+    ///  ▐▐  ▛                                  ┆  
+    ///  ▐▐                                     ┆  
+    ///  ▐▐                                     ┆  
+    ///  ▐▐                                     ┆  
+    ///  ▐▞                                     ┆  
+    ///  ▝                                      ┆  
+    ///  │                                      ┆  
+    ///  │                                      ┆  
+    ///  ───────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    ElasticOut,
+    /// ```txt
+    ///                       ▖                    
+    ///                       █                    
+    ///                       █                    
+    ///  ▲                    █▐▖                  
+    ///  │                    █▐▚▙▗▖▗ ▖            
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▙▌█▐▌█▐▌▛▝▀▀▀▀▀▀▀▀▀▀  
+    ///  │                   █▌█▐▌▛▝            ┆  
+    ///  │                   █▌█▝               ┆  
+    ///  │                   █▌▛                ┆  
+    ///  │                   █▌                 ┆  
+    ///  │                   █▘                 ┆  
+    ///  │                   ▌                  ┆  
+    ///  │                   ▌                  ┆  
+    ///  │                   ▌                  ┆  
+    ///  │                   ▌                  ┆  
+    ///  │                  ▖▌                  ┆  
+    ///  │                  █▌                  ┆  
+    ///  │                  █▌                  ┆  
+    ///  │                ▐▖█▌                  ┆  
+    ///  │          ▖▗ ▄▗▚▌▙▜▌                  ┆  
+    ///  ▀▀▀▀▀▀▀▀▀▀▀▝▞▚▌█▐▌█▐▌──────────────────┬─▶
+    /// 0             ▝ ▛▐▌█▐▌                  1  
+    ///                 ▝ █▐▌                     
+    ///                   ▛▐▌                     
+    ///                    ▐▌                     
+    ///                    ▐▘                     
+    /// ```
+    ElasticInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▐  
+    ///  │                                      ▌  
+    ///  │                                     ▗▘  
+    ///  │                                     ▞┆  
+    ///  │                                    ▗▘┆  
+    ///  │                                    ▞ ┆  
+    ///  │                                   ▗▘ ┆  
+    ///  │                                   ▌  ┆  
+    ///  │                                  ▐   ┆  
+    ///  │                                  ▌   ┆  
+    ///  │                                 ▞    ┆  
+    ///  │                                ▗▘    ┆  
+    ///  │                               ▗▘     ┆  
+    ///  │                               ▞      ┆  
+    ///  ▀▀▀▄▄▖─────────────────────────▞───────┬─▶
+    /// 0     ▝▀▄▖                     ▞        1  
+    ///          ▝▀▄                  ▞            
+    ///             ▀▚▖             ▗▞             
+    ///               ▝▀▄▖         ▄▘              
+    ///                  ▝▀▄▄▄▄▄▄▞▀                
+    /// ```
+    BackIn,
+    /// ```txt
+    ///                ▗▄▀▀▀▀▀▀▚▄                  
+    ///               ▞▘         ▀▚▄               
+    ///             ▗▀              ▀▄▖            
+    ///  ▲         ▗▘                 ▝▚▄          
+    ///  │        ▗▘                     ▀▚▄       
+    /// 1┤┄┄┄┄┄┄┄▗▘┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▀▀▚▄▄  
+    ///  │       ▌                              ┆  
+    ///  │      ▞                               ┆  
+    ///  │     ▐                                ┆  
+    ///  │    ▗▘                                ┆  
+    ///  │    ▞                                 ┆  
+    ///  │   ▗▘                                 ┆  
+    ///  │   ▞                                  ┆  
+    ///  │  ▐                                   ┆  
+    ///  │  ▌                                   ┆  
+    ///  │ ▐                                    ┆  
+    ///  │ ▌                                    ┆  
+    ///  │▐                                     ┆  
+    ///  │▞                                     ┆  
+    ///  ▗▘                                     ┆  
+    ///  ▞──────────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    BackOut,
+    /// ```txt
+    ///                              ▗▄▖           
+    ///  ▲                         ▗▀▘ ▝▀▚▄        
+    ///  │                        ▞▘       ▀▄▖     
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▐┄┄┄┄┄┄┄┄┄┄┄▝▀▄▄  
+    ///  │                      ▗▘              ┆  
+    ///  │                      ▌               ┆  
+    ///  │                     ▐                ┆  
+    ///  │                     ▌                ┆  
+    ///  │                    ▐                 ┆  
+    ///  │                    ▌                 ┆  
+    ///  │                   ▐                  ┆  
+    ///  │                   ▌                  ┆  
+    ///  │                  ▐                   ┆  
+    ///  │                  ▌                   ┆  
+    ///  │                 ▐                    ┆  
+    ///  │                 ▌                    ┆  
+    ///  │                ▐                     ┆  
+    ///  │               ▗▘                     ┆  
+    ///  ▄▄▖─────────────▌──────────────────────┬─▶
+    /// 0  ▝▀▄▖         ▞                       1  
+    ///       ▝▚▄     ▗▀                           
+    ///          ▀▀▄▄▀▘                            
+    /// ```
+    BackInOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗▄  
+    ///  │                                   ▄▀▘┆  
+    ///  │                                 ▗▞   ┆  
+    ///  │                                ▗▘    ┆  
+    ///  │                               ▗▘     ┆  
+    ///  │                              ▗▘      ┆  
+    ///  │                              ▌       ┆  
+    ///  │                             ▞        ┆  
+    ///  │                            ▗▘        ┆  
+    ///  │                            ▌         ┆  
+    ///  │                           ▞          ┆  
+    ///  │              ▗▞▀▀▀▄▖     ▗▘          ┆  
+    ///  │             ▞▘     ▝▚    ▞           ┆  
+    ///  │           ▗▞         ▚  ▗▘           ┆  
+    ///  │    ▗▄▀▚▄  ▞           ▚ ▞            ┆  
+    ///  ▄▀▀▚▞▘────▀▞─────────────▚▘────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    BounceIn,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄▗┄┄┄┄┄┄┄┄┄┄┄┄┄┄▖┄┄┄┄┄┄▄┄┄▗  
+    ///  │             ▞▚            ▞▝▚▄ ▗▄▀ ▀▀▘  
+    ///  │            ▗▘ ▚          ▄▘   ▀▘     ┆  
+    ///  │            ▞   ▚        ▞            ┆  
+    ///  │           ▗▘    ▀▄▖   ▄▀             ┆  
+    ///  │           ▞       ▝▀▀▀               ┆  
+    ///  │          ▐                           ┆  
+    ///  │         ▗▘                           ┆  
+    ///  │         ▞                            ┆  
+    ///  │        ▐                             ┆  
+    ///  │       ▗▘                             ┆  
+    ///  │      ▗▘                              ┆  
+    ///  │     ▗▘                               ┆  
+    ///  │    ▄▘                                ┆  
+    ///  │  ▗▞                                  ┆  
+    ///  ▄▄▀▘───────────────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    BounceOut,
+    /// ```txt
+    ///  ▲                                         
+    ///  │                                         
+    /// 1┤┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄┄▗┄┄┄▄▄  
+    ///  │                          ▛▖    ▗▘▀▀▀ ┆  
+    ///  │                         ▐ ▝▚▄▄▄▘     ┆  
+    ///  │                        ▗▘            ┆  
+    ///  │                        ▞             ┆  
+    ///  │                       ▞              ┆  
+    ///  │                      ▄▘              ┆  
+    ///  │                    ▗▞                ┆  
+    ///  │                 ▄▀▀▘                 ┆  
+    ///  │               ▗▞                     ┆  
+    ///  │               ▞                      ┆  
+    ///  │              ▞                       ┆  
+    ///  │             ▗▘                       ┆  
+    ///  │      ▗▄▄▄   ▌                        ┆  
+    ///  │     ▗▘   ▀▖▐                         ┆  
+    ///  ▄▄▞▀▀▚▘─────▝▘─────────────────────────┬─▶
+    /// 0                                       1  
+    /// ```
+    BounceInOut,
+}
+impl EaseFunction {
+    /// Sample a value from this ease function.
+    pub fn sample(&self, v: f32) -> f32 {
+        use EaseFunction::*;
+        match self {
+            Linear => linear(v),
+            QuadraticIn => quadratic_in(v),
+            QuadraticOut => quadratic_out(v),
+            QuadraticInOut => quadratic_in_out(v),
+            CubicIn => cubic_in(v),
+            CubicOut => cubic_out(v),
+            CubicInOut => cubic_in_out(v),
+            QuarticIn => quartic_in(v),
+            QuarticOut => quartic_out(v),
+            QuarticInOut => quartic_in_out(v),
+            QuinticIn => quintic_in(v),
+            QuinticOut => quintic_out(v),
+            QuinticInOut => quintic_in_out(v),
+            SineIn => sine_in(v),
+            SineOut => sine_out(v),
+            SineInOut => sine_in_out(v),
+            CircularIn => circular_in(v),
+            CircularOut => circular_out(v),
+            CircularInOut => circular_in_out(v),
+            ExponentialIn => exponential_in(v),
+            ExponentialOut => exponential_out(v),
+            ExponentialInOut => exponential_in_out(v),
+            ElasticIn => elastic_in(v),
+            ElasticOut => elastic_out(v),
+            ElasticInOut => elastic_in_out(v),
+            BackIn => back_in(v),
+            BackOut => back_out(v),
+            BackInOut => back_in_out(v),
+            BounceIn => bounce_in(v),
+            BounceOut => bounce_out(v),
+            BounceInOut => bounce_in_out(v),
+        }
+    }
+}
+
+#[allow(clippy::type_complexity)]
+pub fn ease_function_a_to_b_system<V>(
+    f: impl Send + Sync + 'static + Fn(&V, &V, &f32) -> V,
+) -> impl Fn(
+    Commands,
+    Query<
+        (Entity, &AToB<V, EaseFunction>, &TimeSpanProgress),
+        Or<(Changed<EaseFunction>, Changed<TimeSpanProgress>)>,
+    >,
+    RemovedComponents<TimeSpanProgress>,
+)
+where
+    V: Send + Sync + 'static,
+{
+    move |mut commands, query, mut removed| {
+        query.iter().for_each(|(entity, a_to_b, progress)| {
+            if progress.now_percentage.is_nan() {
+                return;
+            }
+            let value = a_to_b.curve.sample(progress.now_percentage);
+
+            commands
+                .entity(entity)
+                .insert(CurveValue(f(&a_to_b.a, &a_to_b.b, &value)));
+        });
+        removed.read().for_each(|entity| {
+            if let Some(mut entity) = commands.get_entity(entity) {
+                entity.remove::<CurveValue<V>>();
+            }
+        });
+    }
+}
+
 fn clamp(p: f32) -> f32 {
     match () {
         _ if p > 1.0 => 1.0,
@@ -7,21 +831,21 @@ fn clamp(p: f32) -> f32 {
     }
 }
 
-pub fn linear(v: f32) -> f32 {
+fn linear(v: f32) -> f32 {
     clamp(v)
 }
 
-pub fn quadratic_in(v: f32) -> f32 {
+fn quadratic_in(v: f32) -> f32 {
     let p = clamp(v);
     p * p
 }
 
-pub fn quadratic_out(v: f32) -> f32 {
+fn quadratic_out(v: f32) -> f32 {
     let p = clamp(v);
     -(p * (p - 2.0))
 }
 
-pub fn quadratic_in_out(v: f32) -> f32 {
+fn quadratic_in_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 0.5 {
         2.0 * p * p
@@ -30,18 +854,18 @@ pub fn quadratic_in_out(v: f32) -> f32 {
     }
 }
 
-pub fn cubic_in(v: f32) -> f32 {
+fn cubic_in(v: f32) -> f32 {
     let p = clamp(v);
     p * p * p
 }
 
-pub fn cubic_out(v: f32) -> f32 {
+fn cubic_out(v: f32) -> f32 {
     let p = clamp(v);
     let f = p - 1.0;
     f * f * f + 1.0
 }
 
-pub fn cubic_in_out(v: f32) -> f32 {
+fn cubic_in_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 0.5 {
         4.0 * p * p * p
@@ -51,18 +875,18 @@ pub fn cubic_in_out(v: f32) -> f32 {
     }
 }
 
-pub fn quartic_in(v: f32) -> f32 {
+fn quartic_in(v: f32) -> f32 {
     let p = clamp(v);
     p * p * p * p
 }
 
-pub fn quartic_out(v: f32) -> f32 {
+fn quartic_out(v: f32) -> f32 {
     let p = clamp(v);
     let f = p - 1.0;
     f * f * f * (1.0 - p) + 1.0
 }
 
-pub fn quartic_in_out(v: f32) -> f32 {
+fn quartic_in_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 0.5 {
         8.0 * p * p * p * p
@@ -72,18 +896,18 @@ pub fn quartic_in_out(v: f32) -> f32 {
     }
 }
 
-pub fn quintic_in(v: f32) -> f32 {
+fn quintic_in(v: f32) -> f32 {
     let p = clamp(v);
     p * p * p * p * p
 }
 
-pub fn quintic_out(v: f32) -> f32 {
+fn quintic_out(v: f32) -> f32 {
     let p = clamp(v);
     let f = p - 1.0;
     f * f * f * f * f + 1.0
 }
 
-pub fn quintic_in_out(v: f32) -> f32 {
+fn quintic_in_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 0.5 {
         16.0 * p * p * p * p * p
@@ -93,32 +917,32 @@ pub fn quintic_in_out(v: f32) -> f32 {
     }
 }
 
-pub fn sine_in(v: f32) -> f32 {
+fn sine_in(v: f32) -> f32 {
     let p = clamp(v);
     1. - (p * PI * 0.5).cos()
 }
 
-pub fn sine_out(v: f32) -> f32 {
+fn sine_out(v: f32) -> f32 {
     let p = clamp(v);
     (p * PI * 0.5).sin()
 }
 
-pub fn sine_in_out(v: f32) -> f32 {
+fn sine_in_out(v: f32) -> f32 {
     let p = clamp(v);
     -((p * PI).cos() - 1.) * 0.5
 }
 
-pub fn circular_in(v: f32) -> f32 {
+fn circular_in(v: f32) -> f32 {
     let p = clamp(v);
     1.0 - (1.0 - (p * p)).sqrt()
 }
 
-pub fn circular_out(v: f32) -> f32 {
+fn circular_out(v: f32) -> f32 {
     let p = clamp(v);
     ((2.0 - p) * p).sqrt()
 }
 
-pub fn circular_in_out(v: f32) -> f32 {
+fn circular_in_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 0.5 {
         0.5 * (1.0 - (1.0 - 4.0 * (p * p)).sqrt())
@@ -127,7 +951,7 @@ pub fn circular_in_out(v: f32) -> f32 {
     }
 }
 
-pub fn exponential_in(v: f32) -> f32 {
+fn exponential_in(v: f32) -> f32 {
     if v <= 0.0 {
         0.0
     } else {
@@ -135,7 +959,7 @@ pub fn exponential_in(v: f32) -> f32 {
     }
 }
 
-pub fn exponential_out(v: f32) -> f32 {
+fn exponential_out(v: f32) -> f32 {
     if v >= 1.0 {
         1.0
     } else {
@@ -143,7 +967,7 @@ pub fn exponential_out(v: f32) -> f32 {
     }
 }
 
-pub fn exponential_in_out(v: f32) -> f32 {
+fn exponential_in_out(v: f32) -> f32 {
     if v <= 0.0 {
         return 0.0;
     }
@@ -158,42 +982,40 @@ pub fn exponential_in_out(v: f32) -> f32 {
     }
 }
 
-pub fn elastic_in(v: f32) -> f32 {
+fn elastic_in(v: f32) -> f32 {
     let p = clamp(v);
-    (13.0 * std::f32::consts::TAU * p).sin() * (2.0_f32).powf(10.0 * (p - 1.0))
+    (13.0 * TAU * p).sin() * (2.0_f32).powf(10.0 * (p - 1.0))
 }
 
-pub fn elastic_out(v: f32) -> f32 {
+fn elastic_out(v: f32) -> f32 {
     let p = clamp(v);
-    (-13.0 * std::f32::consts::TAU * (p + 1.0)).sin()
-        * (2.0_f32).powf(-10.0 * p)
-        + 1.0
+    (-13.0 * TAU * (p + 1.0)).sin() * (2.0_f32).powf(-10.0 * p) + 1.0
 }
 
-pub fn elastic_in_out(v: f32) -> f32 {
+fn elastic_in_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 0.5 {
-        0.5 * (13.0 * std::f32::consts::TAU * (2.0 * p)).sin()
+        0.5 * (13.0 * TAU * (2.0 * p)).sin()
             * (2.0_f32).powf(10.0 * ((2.0 * p) - 1.0))
     } else {
-        0.5 * ((-13.0 * std::f32::consts::TAU * ((2.0 * p - 1.0) + 1.0)).sin()
+        0.5 * ((-13.0 * TAU * ((2.0 * p - 1.0) + 1.0)).sin()
             * (2.0_f32).powf(-10.0 * (2.0 * p - 1.0))
             + 2.0)
     }
 }
 
-pub fn back_in(v: f32) -> f32 {
+fn back_in(v: f32) -> f32 {
     let p = clamp(v);
     p * p * p - p * (p * PI).sin()
 }
 
-pub fn back_out(v: f32) -> f32 {
+fn back_out(v: f32) -> f32 {
     let p = clamp(v);
     let f = 1.0 - p;
     1.0 - (f * f * f - f * (f * PI).sin())
 }
 
-pub fn back_in_out(v: f32) -> f32 {
+fn back_in_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 0.5 {
         let f = 2.0 * p;
@@ -204,12 +1026,12 @@ pub fn back_in_out(v: f32) -> f32 {
     }
 }
 
-pub fn bounce_in(v: f32) -> f32 {
+fn bounce_in(v: f32) -> f32 {
     let p = clamp(v);
     1.0 - bounce_out(1.0 - p)
 }
 
-pub fn bounce_out(v: f32) -> f32 {
+fn bounce_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 4.0 / 11.0 {
         (121.0 * p * p) / 16.0
@@ -222,7 +1044,7 @@ pub fn bounce_out(v: f32) -> f32 {
     }
 }
 
-pub fn bounce_in_out(v: f32) -> f32 {
+fn bounce_in_out(v: f32) -> f32 {
     let p = clamp(v);
     if p < 0.5 {
         0.5 * bounce_in(p * 2.0)
